Remove remaining test warnings across CI environments
Test fixes:
- test_units: fill the missing-data datasets with a float dtype instead
  of casting NaN into int64 arrays.
- test_FillValueCoder_complex_roundtrip: draw values representable in
  the target complex dtype instead of overflowing when casting.
- TestReduction.test_mean: ignore floating point overflow/invalid
  warnings from arbitrary hypothesis floats.
- test_show_versions: ignore setuptools' distutils warning, raised by
  importing all dependencies.

Third-party warnings:
- Fix the category of the pydap/webob "cgi" import filter
  (DeprecationWarning, not FutureWarning).
- Add google-crc32c to the min-versions environment; numcodecs warns
  on import with crc32c and overrides any warning filter.
- Ignore zarr < 3.1 "not part of the Zarr format 3 specification"
  warnings, netcdf4-python < 1.7 empty array truth-value warnings, and
  (with numpy < 2) h5netcdf's array-to-scalar conversion warnings.

+++ xarray/tests/test_units.py
@@ -864,8 +864,14 @@ def test_combine_nested(variant, unit, error, dtype):
     )
     ds3 = xr.Dataset(
         data_vars={
-            "a": (("y", "x"), np.full_like(array1, fill_value=np.nan) * data_unit2),
-            "b": (("y", "x"), np.full_like(array2, fill_value=np.nan) * data_unit2),
+            "a": (
+                ("y", "x"),
+                np.full_like(array1, fill_value=np.nan, dtype=np.float64) * data_unit2,
+            ),
+            "b": (
+                ("y", "x"),
+                np.full_like(array2, fill_value=np.nan, dtype=np.float64) * data_unit2,
+            ),
         },
         coords={
             "x": np.arange(3, 6) * dim_unit2,
@@ -1209,8 +1215,14 @@ def test_merge_dataset(variant, unit, error, dtype):
     )
     ds3 = xr.Dataset(
         data_vars={
-            "a": (("y", "x"), np.full_like(array1, np.nan) * data_unit2),
-            "b": (("y", "x"), np.full_like(array2, np.nan) * data_unit2),
+            "a": (
+                ("y", "x"),
+                np.full_like(array1, np.nan, dtype=np.float64) * data_unit2,
+            ),
+            "b": (
+                ("y", "x"),
+                np.full_like(array2, np.nan, dtype=np.float64) * data_unit2,
+            ),
         },
         coords={
             "x": np.arange(3, 6) * dim_unit2,
